Product Variant Item Done

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -94,10 +94,12 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $varinat = ProductVariant::findOrFail($id);
-        // $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
-        // if($variantItemCheck > 0){
-        //     return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
-        // }
+        
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
+        if($variantItemCheck > 0){
+            return response(['status' => 'error', 'message' => 'This variant contain variant items in it. Delete the variant items first to delete this variant!']);
+        }
+        
         $varinat->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
